feat: quality scale self-assessed Platinum — translated action errors, strict typing for the websocket package
- exception-translations (the last open Gold rule): all user-facing raises
  in the actions, the task button and the skip lock carry translation keys
  (11 new keys, strings.json + 21 translations); only internal codes the WS
  layer maps itself stay untranslated. tests/test_exception_translations.py
  tripwires code ↔ strings.json (key exists, placeholders match, no unused
  string, no new untranslated user-facing raise).
- strict-typing: the websocket package is no longer excluded from mypy
  --strict; only homeassistant.components.websocket_api gets
  implicit_reexport. 39 real type errors fixed without behaviour change
  (entry narrowing, foreign_part_resolver annotations, Final fields across
  the import cycle, part stock accepts fractional absolute values as the
  schema already did).
- entity-category: the document-storage sensor is diagnostic.
- quality_scale.yaml brought up to date (pypdf dependency, 22 languages,
  4,200+ tests, docs-supported-devices done via INTEGRATIONS.md); manifest
  and README say Platinum; ARCHITECTURE points to the yaml plus the guards.

+++ custom_components/maintenance_supporter/parts_runtime.py
@@ -124,7 +124,7 @@ async def async_change_part_stock(
     part_id: str,
     *,
     delta: float | None = None,
-    absolute: int | None = None,
+    absolute: float | None = None,
 ) -> float | None:
     """Change one part's stock (clamped at 0), fire the edge event, save.
